+++ demos/cypress-shared/cypress/e2e/correct.captcha.cy.ts
@@ -16,24 +16,23 @@ import "@cypress/xpath";
 import { u8aToHex } from "@polkadot/util";
 import { ProsopoDatasetError } from "@prosopo/common";
 import { datasetWithSolutionHashes } from "@prosopo/datasets";
-import { getPairAsync } from "@prosopo/keyring";
+import { getPair } from "@prosopo/keyring";
 import {
 	AdminApiPaths,
 	type Captcha,
-	CaptchaType,
+	type CaptchaType,
 	type IUserSettings,
 	type RegisterSitekeyBodyTypeOutput,
 	Tier,
 } from "@prosopo/types";
 import { checkboxClass, getWidgetElement } from "../support/commands.js";
 
-let captchaType: CaptchaType;
+const baseCaptchaType: CaptchaType = Cypress.env("CAPTCHA_TYPE") || "image";
 
 describe("Captchas", () => {
 	before(() => {
-		captchaType = Cypress.env("CAPTCHA_TYPE") || "image";
 		// Call registerSiteKey and handle response here
-		return cy.registerSiteKey(captchaType).then((response) => {
+		return cy.registerSiteKey(baseCaptchaType).then((response) => {
 			// Log the response status and body using cy.task()
 			cy.task("log", `Response status: ${response.status}`);
 			cy.task("log", `Response: ${JSON.stringify(response.body)}`);
@@ -76,7 +75,7 @@ describe("Captchas", () => {
 	});
 
 	after(() => {
-		return cy.registerSiteKey(CaptchaType.image);
+		return cy.registerSiteKey(baseCaptchaType);
 	});
 
 	it("Selecting the incorrect images fails the captcha", () => {