+++ packages/cli/src/reloader.ts
@@ -13,10 +13,10 @@
 // limitations under the License.
 import fs from "node:fs";
 import type { Server } from "node:net";
-import type { KeyringPair } from "@polkadot/keyring/types";
 import { LogLevel, getLogger } from "@prosopo/common";
 import { loadEnv } from "@prosopo/dotenv";
 import { ProviderEnvironment } from "@prosopo/env";
+import type { KeyringPair } from "@prosopo/types";
 import type { ProsopoConfigOutput } from "@prosopo/types";
 import type { AwaitedProcessedArgs } from "./argv.js";
 import { start, startDev } from "./start.js";
@@ -55,7 +55,7 @@ export default class ReloadingAPI {
 	}
 
 	public async start(reloadEnv = false) {
-		log.info("Starting API");
+		log.info(() => ({ msg: "Starting API" }));
 		this._envWatcher = await this._watchEnv();
 		loadEnv();
 		if (!this._env && reloadEnv) {
@@ -71,7 +71,7 @@ export default class ReloadingAPI {
 	}
 
 	public async startDev() {
-		log.info("Starting API");
+		log.info(() => ({ msg: "Starting API" }));
 		this._envWatcher = await this._watchEnv();
 		loadEnv();
 		const env = new ProviderEnvironment(
@@ -84,7 +84,7 @@ export default class ReloadingAPI {
 	}
 
 	public async stop() {
-		log.info("Stopping API");
+		log.info(() => ({ msg: "Stopping API" }));
 		return new Promise((resolve) => {
 			if (this.api) {
 				this.api.close(resolve);
@@ -94,7 +94,10 @@ export default class ReloadingAPI {
 
 	private async _watchEnv() {
 		return fs.watchFile(this._envPath, async () => {
-			log.info(`env file change detected. Restarting: ${this._restarting}`);
+			log.info(() => ({
+				data: { restarting: this._restarting },
+				msg: "env file change detected. Restarting",
+			}));
 			if (!this._restarting) {
 				this._restarting = true;
 				await this.stop();