+++ packages/types/src/provider/api.ts
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import type { ApiJsonError } from "@prosopo/common";
+import type { ApiJsonError, LogObject } from "@prosopo/common";
 import type { Address4, Address6 } from "ip-address";
 import {
 	type ZodDefault,
@@ -72,6 +72,7 @@ export enum ClientApiPaths {
 }
 
 export enum PublicApiPaths {
+	Healthz = "/healthz",
 	GetProviderDetails = "/v1/prosopo/provider/public/details",
 }