+++ dev/config/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @prosopo/config
 
+## 3.0.0
+### Major Changes
+
+- 64b5bcd: Access Controls
+
+## 2.6.1
+### Patch Changes
+
+- 86c22b8: structured logging
+
 ## 2.6.0
 
 ### Minor Changes