+++ dev/lint/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @prosopo/lint
 
+## 2.7.2
+### Patch Changes
+
+- Updated dependencies [64b5bcd]
+  - @prosopo/util@3.0.0
+
+## 2.7.1
+### Patch Changes
+
+  - @prosopo/util@2.6.1
+
+## 2.7.0
+### Minor Changes
+
+- 8f0644a: Taking required functions from polkadot/keyring and polkadot/util-crypto in-house and removing WASM dependencies. Adding @scure JS-based sr25519 function instead.
+
 ## 2.6.0
 
 ### Minor Changes