+++ packages/api-route/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @prosopo/api-route
 
+## 2.6.5
+### Patch Changes
+
+- Updated dependencies [64b5bcd]
+  - @prosopo/common@3.0.0
+
+## 2.6.4
+### Patch Changes
+
+- 86c22b8: structured logging
+- Updated dependencies [86c22b8]
+  - @prosopo/common@2.7.2
+
+## 2.6.3
+### Patch Changes
+
+  - @prosopo/common@2.7.1
+
+## 2.6.2
+### Patch Changes
+
+- Updated dependencies [8f0644a]
+  - @prosopo/common@2.7.0
+
 ## 2.6.1
 
 ### Patch Changes